Rewrite utils function
