+++ app/changelog/page.tsx
@@ -1,32 +1,13 @@
 import {Fragment} from 'react';
 import * as Sentry from '@sentry/nextjs';
 import type {Metadata} from 'next';
-import {unstable_cache} from 'next/cache';
 
 import List from 'sentry-docs/components/changelog/list';
-import prisma from 'sentry-docs/prisma';
 
 import Header from './header';
+import {getChangelogs} from './utils';
 
 export const dynamic = 'force-dynamic';
-
-export const getChangelogs = unstable_cache(
-  async () => {
-    return await prisma.changelog.findMany({
-      include: {
-        categories: true,
-      },
-      where: {
-        published: true,
-      },
-      orderBy: {
-        publishedAt: 'desc',
-      },
-    });
-  },
-  ['changelogs'],
-  {tags: ['changelogs']}
-);
 
 export default async function ChangelogList() {
   const changelogs = await getChangelogs();
